add functions and tests

+++ src/main/java/fr/pala/accounting/dal/AccountDAL.java
@@ -14,7 +14,7 @@ public interface AccountDAL {
 
     AccountModel getAccountOfUser(String user_id, String account_id);
 
-    Float getAmountOfAccount(String user_id,String account_id);
+    Double getAmountOfAccount(String user_id,String account_id);
 
     void updateAccount(String user_id, String account_id, AccountModel account);
